move quicktype options to own file
